backend/django_services/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
crewai==0.70.1
crewai-tools==0.12.1
openai>=1.13.3,<2.0.0
python-dotenv==1.0.0
requests==2.31.0
